Move query to common definitions and add openapi-types to dev dependencies
