+++ packages/server/package.json
@@ -111,7 +111,6 @@
     "mysql2": "^2.3.1",
     "node-fetch": "2.6.0",
     "open": "^8.4.0",
-    "openapi-types": "^9.3.1",
     "pg": "8.5.1",
     "pino-pretty": "4.0.0",
     "posthog-node": "^1.1.4",
@@ -150,6 +149,7 @@
     "eslint": "^6.8.0",
     "jest": "^27.0.5",
     "nodemon": "^2.0.4",
+    "openapi-types": "^9.3.1",
     "path-to-regexp": "^6.2.0",
     "prettier": "^2.3.1",
     "rimraf": "^3.0.2",
