+++ packages/server/src/api/controllers/query/import/sources/curl.ts
@@ -1,4 +1,5 @@
-import { ImportSource, ImportInfo, Query } from "./base"
+import { ImportSource, ImportInfo } from "./base"
+import { Query } from "../../../../../definitions/common"
 import { URL } from "url"
 const curlconverter = require("curlconverter")
