+++ packages/server/src/definitions/common.ts
@@ -107,3 +107,25 @@ export interface Datasource extends Base {
     [key: string]: Table
   }
 }
+
+export interface QueryParameter {
+  name: string
+  default: string
+}
+
+export interface Query {
+  _id?: string
+  datasourceId: string
+  name: string
+  parameters: QueryParameter[]
+  fields: {
+    headers: object
+    queryString: string | null
+    path: string
+    requestBody: string | undefined
+  }
+  transformer: string | null
+  schema: any
+  readable: boolean
+  queryVerb: string
+}
